# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Club {
  id: number;
  name: string;
  shortName: string;
  category: string;
  tagline: string;
  officeLocation: string;
  memberCount: number;
  accent: string;
  isVerified: boolean;
  isSaved?: boolean;
}

export interface ClubRole {
  id: number;
  title: string;
  responsibilities: string[];
  skills: string[];
  openings: number;
}

export type ClubDetail = Club & {
  description: string;
  eligibility: string;
  roles: ClubRole[];
  meetingSchedule: string;
  socials: string[];
  joinedCount: number;
};

export interface Event {
  id: number;
  title: string;
  clubId: number;
  clubName: string;
  category: string;
  startsAt: string;
  endsAt: string;
  venue: string;
  entryFee: number;
  dutyLeave: boolean;
  spotsLeft: number;
  registrationRequired: boolean;
  accent: string;
  description?: string;
}

export type EventDetail = Event & {
  registrationCriteria: string;
  organizerEmail: string;
  registeredCount: number;
};

export interface RegistrationInput {
  /** @minLength 2 */
  studentName: string;
  studentEmail: string;
  year: string;
  department: string;
}

export interface Registration {
  id: number;
  eventId: number;
  studentName: string;
  studentEmail: string;
  status: string;
  createdAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface Dashboard {
  clubCount: number;
  eventCount: number;
  savedCount: number;
  thisWeekEvents: number;
  categories: CategoryCount[];
  featuredClubs: Club[];
  nextEvent: Event | null;
}

export type ListClubsParams = {
search?: string;
category?: string;
sort?: ListClubsSort;
};

export type ListClubsSort = typeof ListClubsSort[keyof typeof ListClubsSort];


export const ListClubsSort = {
  popular: 'popular',
  alphabetical: 'alphabetical',
  active: 'active',
} as const;

export type ListEventsParams = {
search?: string;
category?: string;
timeframe?: ListEventsTimeframe;
};

export type ListEventsTimeframe = typeof ListEventsTimeframe[keyof typeof ListEventsTimeframe];


export const ListEventsTimeframe = {
  week: 'week',
  month: 'month',
  all: 'all',
} as const;