# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: number;
  name: string;
  shortName: string;
  category: string;
  tagline: string;
  officeLocation: string;
  memberCount: number;
  accent: string;
  isVerified: boolean;
  isSaved?: boolean;
}