# lib/api-zod/src/generated/types/clubDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */
import type { Club } from './club';
import type { ClubRole } from './clubRole';

export type ClubDetail = Club & {
  description: string;
  eligibility: string;
  roles: ClubRole[];
  meetingSchedule: string;
  socials: string[];
  joinedCount: number;
};