# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { Club } from './club';
import type { Event } from './event';

export interface Dashboard {
  clubCount: number;
  eventCount: number;
  savedCount: number;
  thisWeekEvents: number;
  categories: CategoryCount[];
  featuredClubs: Club[];
  nextEvent: Event | null;
}