# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  clubId: number;
  clubName: string;
  category: string;
  startsAt: Date;
  endsAt: Date;
  venue: string;
  entryFee: number;
  dutyLeave: boolean;
  spotsLeft: number;
  registrationRequired: boolean;
  accent: string;
  description?: string;
}