# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Clubs Hub student directory and events platform
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 2 */
  studentName: string;
  studentEmail: string;
  year: string;
  department: string;
}